Move controller factory to POS and make POS-oriented
