+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/woopos/home/totals/WooPosTotalsViewModelTest.kt
@@ -21,7 +21,6 @@ import com.woocommerce.android.ui.payments.cardreader.payment.CardReaderPaymentE
 import com.woocommerce.android.ui.payments.cardreader.payment.CardReaderPaymentOrderHelper
 import com.woocommerce.android.ui.payments.cardreader.payment.PaymentFlowError
 import com.woocommerce.android.ui.payments.cardreader.payment.controller.CardReaderPaymentController
-import com.woocommerce.android.ui.payments.cardreader.payment.controller.CardReaderPaymentControllerFactory
 import com.woocommerce.android.ui.payments.cardreader.payment.controller.CardReaderPaymentOrRefundState
 import com.woocommerce.android.ui.payments.cardreader.payment.controller.CardReaderPaymentOrRefundState.CardReaderPaymentState
 import com.woocommerce.android.ui.payments.cardreader.payment.controller.CardReaderPaymentStateProvider
@@ -102,7 +101,7 @@ class WooPosTotalsViewModelTest {
     private val cardReaderConfigProvider: CardReaderCountryConfigProvider = mock()
     private val paymentReceiptShare: PaymentReceiptShare = mock()
     private val uiStringParser: UiStringParser = mock()
-    private val paymentControllerFactory = CardReaderPaymentControllerFactory(
+    private val paymentControllerFactory = WooPosCardReaderPaymentControllerFactory(
         cardReaderManager = cardReaderManager,
         orderRepository = orderRepository,
         selectedSite = selectedSite,
@@ -562,7 +561,7 @@ class WooPosTotalsViewModelTest {
         whenever(uiStringParser.asString(any())).thenReturn("Unfortunately, this payment has been declined.")
 
         val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-        val factory: CardReaderPaymentControllerFactory = mock()
+        val factory: WooPosCardReaderPaymentControllerFactory = mock()
         whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
         val paymentState =
             MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -697,7 +696,7 @@ class WooPosTotalsViewModelTest {
         whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
 
         val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-        val factory: CardReaderPaymentControllerFactory = mock()
+        val factory: WooPosCardReaderPaymentControllerFactory = mock()
         whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
         createViewModelAndSetupForSuccessfulOrderCreation(controllerFactory = factory)
 
@@ -715,7 +714,7 @@ class WooPosTotalsViewModelTest {
         whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
 
         val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-        val factory: CardReaderPaymentControllerFactory = mock()
+        val factory: WooPosCardReaderPaymentControllerFactory = mock()
         whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
         val vm = createViewModelAndSetupForSuccessfulOrderCreation(controllerFactory = factory)
 
@@ -745,7 +744,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -779,7 +778,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -815,7 +814,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -844,7 +843,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -883,7 +882,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -925,7 +924,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -968,7 +967,7 @@ class WooPosTotalsViewModelTest {
         val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
         whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
         val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-        val factory: CardReaderPaymentControllerFactory = mock()
+        val factory: WooPosCardReaderPaymentControllerFactory = mock()
         whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
         val paymentState =
             MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -1006,7 +1005,7 @@ class WooPosTotalsViewModelTest {
             val readerStatus = MutableStateFlow<CardReaderStatus>(CardReaderStatus.Connected(mock()))
             whenever(cardReaderFacade.readerStatus).thenReturn(readerStatus)
             val mockCardReaderPaymentController: CardReaderPaymentController = mock()
-            val factory: CardReaderPaymentControllerFactory = mock()
+            val factory: WooPosCardReaderPaymentControllerFactory = mock()
             whenever(factory.create(any(), any(), any())).thenReturn(mockCardReaderPaymentController)
             val paymentState =
                 MutableStateFlow<CardReaderPaymentOrRefundState>(
@@ -1147,7 +1146,7 @@ class WooPosTotalsViewModelTest {
 
     @Suppress("LongMethod")
     private suspend fun createViewModelAndSetupForSuccessfulOrderCreation(
-        controllerFactory: CardReaderPaymentControllerFactory = paymentControllerFactory,
+        controllerFactory: WooPosCardReaderPaymentControllerFactory = paymentControllerFactory,
         itemClickedData: List<WooPosItemsViewModel.ItemClickedData> = listOf(
             WooPosItemsViewModel.ItemClickedData.SimpleProduct(
                 id = 1L
@@ -1235,7 +1234,7 @@ class WooPosTotalsViewModelTest {
         totalsRepository: WooPosTotalsRepository = mock(),
         priceFormat: WooPosFormatPrice = mock(),
         savedState: SavedStateHandle = SavedStateHandle(),
-        cardReaderPaymentControllerFactory: CardReaderPaymentControllerFactory = paymentControllerFactory,
+        cardReaderPaymentControllerFactory: WooPosCardReaderPaymentControllerFactory = paymentControllerFactory,
     ) = WooPosTotalsViewModel(
         resourceProvider = resourceProvider,
         parentToChildrenEventReceiver = parentToChildrenEventReceiver,
